[refactor] Create GameViewModel
